Replace the code completion example with an emoji completion demo
Type `:smile` in the new egui demo to complete emoji names.
The kittest tests move to `tests/egui_tests`.
