+++ crates/egui_demo_lib/src/demo/mod.rs
@@ -7,6 +7,7 @@
 pub mod about;
 pub mod code_editor;
 pub mod code_example;
+pub mod completion;
 pub mod dancing_strings;
 pub mod demo_app_windows;
 pub mod drag_and_drop;
