+++ crates/egui_demo_lib/tests/snapshots/demos/Completion.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:05e098c9741cce3ddf4b88930394cb5114337e7988d3db643f5d815c522e0616
+size 24936
