--- examples/code_completion/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "code_completion"
-edition.workspace = true
-license.workspace = true
-publish = false
-rust-version.workspace = true
-version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-eframe = { workspace = true, features = [
-  "default",
-  "__screenshot", # __screenshot is so we can dump a screenshot using EFRAME_SCREENSHOT_TO
-] }
-env_logger = { workspace = true, features = ["auto-color", "humantime"] }
-
-
-[lints]
-workspace = true
-
-[dev-dependencies]
-egui_kittest = { workspace = true }
