--- examples/code_completion/README.md
@@ -1,9 +1,0 @@
-Example of a code completion popup over a `TextEdit`.
-
-Type `/` to get a list of matching commands. Use the arrow keys to select one, and Enter or Tab to accept it. Escape closes the popup. Keyboard focus stays in the text field the whole time.
-
-```sh
-cargo run -p code_completion
-```
-
-![](screenshot.png)
