--- examples/code_completion/screenshot.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:3a14075e192294a5be5ff784b45616f98ecaca526e9f330ebe8e2b83b8443b50
-size 42565
